Seperated modal from card, for making it a functional component
